Improve average for illumination correction
Add option to do the average by batch and take the median of the batch.
This keep the memory needed manageable and should remove isolated very
bright spots.

Add simple filter option to exclude specific tiffs. Only in subfunctions.
No need to expose to CLI as it needs to be done manually for tiff selection
anyways.