[bugfix] add indication of missing data shape
correct_shifts was used for fluorescence tile, where it register 4 channels together and for
reg2ref where it register 1 channel
Missing data shape must be specified if we want to use the same function.
New argument for that. If it is None (default) crahs with missing data (useful to hunt for failed job)